rename checkOrigin.js
